refactor: ROUTE route.ts에서 service.ts로 이동
