+++ src/__tests__/components/app/SignIn.test.tsx
@@ -5,7 +5,7 @@ import type {TProps as TSecretInputProps} from "@/components/SecretInput"
 import Context from "@/store/sign-in"
 import type {TProps as TPortalProps} from "@/components/Portal"
 import {SIGN_IN} from "@/constants/response-code"
-import ROUTE from "@/constants/route"
+import {ROUTE} from "@/constants/service"
 import {createLocalStorageMock} from "@/__mocks__/window"
 import type {TRoute, TTab} from "@/@types/sign-in"
